Added functionality so that the user can select what size area, small medium and large.a area is aswell now created to the size of the users selection on start.

+++ cs4227/src/CreateMaze/CoinTile.java
@@ -25,9 +25,9 @@ public class CoinTile extends Coin
     }
 
     @Override
-    public double damage() 
+    public int damage() 
     {
-        return 0.0;
+        return 0;
     }
     
 }
