+++ cs4227/src/CreateMaze/Part.java
@@ -9,5 +9,5 @@ public interface Part
 {
     public String name();
     public Boolean destructable();
-    public double damage();
+    public int damage();
 }
